+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ### 2025-06-06
+- [Patch v5.9.15] Add wfv unit tests
+- New/Updated unit tests added for tests.test_wfv_full
+- QA: pytest -q passed (11 tests)
 - [Patch v5.9.13] Improve coverage to 70%
 - New/Updated unit tests added for tests.test_signal_classifier_additional
 - QA: pytest -q passed (698 tests)